third_party/blink/web_tests/platform/generic/fast/dom/HTMLAnchorElement/remove-and-add-href-from-focused-anchor-expected.txt

PASS document.activeElement is target
Remove href.
PASS document.activeElement is target
PASS successfullyParsed is true

TEST COMPLETE
